pyproject.toml
[tool.poetry]
name = "projectonlineschool"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
psycopg2 = "^2.9.10"
pillow = "^11.2.1"
dotenv = "^0.9.9"
djangorestframework = "^3.16.0"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
